TestCaseGenerator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
tenacity==8.2.3
